+++ src/main/java/Handlers/TaskManager.java
@@ -1,4 +1,7 @@
-package Tasks;
+package Handlers;
+
+import Tasks.Task;
+
 public class TaskManager {
     private static Task[] taskList = new Task[100];
     private static int taskCount = 0;
